refactor: consolidate agent hook contracts
