+++ src/bub/builtin/model_runner.py
@@ -25,12 +25,20 @@ from any_llm.types.completion import (
 from loguru import logger
 from pydantic import TypeAdapter, ValidationError
 
-from bub.agent_hooks import LlmCallDecision, LlmCallRequest, LlmCallResult
 from bub.builtin.codex_provider import OpenaiCodexProvider, should_use_openai_codex_provider
 from bub.builtin.settings import AgentSettings, ModelCandidate
 from bub.builtin.tape import Tape
 from bub.hook_runtime import AgentHooks
-from bub.runtime import AsyncStreamEvents, BubError, ErrorKind, StreamEvent, StreamState
+from bub.runtime import (
+    AsyncStreamEvents,
+    BubError,
+    ErrorKind,
+    LlmCallDecision,
+    LlmCallRequest,
+    LlmCallResult,
+    StreamEvent,
+    StreamState,
+)
 from bub.tools import Tool, ToolContext, ToolExecutor
 
 CONTEXT_LENGTH_PATTERNS = re.compile(
