+++ src/bub/hookspecs.py
@@ -7,15 +7,16 @@ from typing import TYPE_CHECKING, Any
 
 import pluggy
 
-from bub.agent_hooks import (
+from bub.runtime import (
+    AsyncStreamEvents,
     LlmCallDecision,
     LlmCallRequest,
     LlmCallResult,
+    RuntimeOptions,
     ToolCall,
     ToolCallDecision,
     ToolCallResult,
 )
-from bub.runtime import AsyncStreamEvents, RuntimeOptions
 from bub.tape import AsyncTapeStore, TapeContext, TapeStore
 from bub.turn_admission import AdmitDecision, TurnSnapshot
 from bub.types import Envelope, MessageHandler, State, SteeringInboxProtocol
